+++ yash-env/CHANGELOG.md
@@ -13,6 +13,12 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 ### Added
 
+- The `system::FdSet` trait has been added to represent a set of file
+  descriptors, abstracting the `fd_set` type from POSIX.
+- The `system::real::FdSet` struct has been added as the concrete `FdSet`
+  implementation for the real system.
+- The `system::virtual::fd_set` module has been added, containing the `FdSet`
+  struct as the concrete `FdSet` implementation for the virtual system.
 - The `system::Sigset` trait has been added to represent a set of signals,
   abstracting the `sigset_t` type from POSIX.
 - The `system::real::Sigset` struct has been added as the concrete `Sigset`
@@ -107,6 +113,9 @@ A _private dependency_ is used internally and not visible to downstream users.
     - `system::Select::select`
 - The `system::Select` trait now requires the `system::Sigmask` trait as a
   supertrait.
+- The `system::Select` trait now has the associated type `FdSet`, which must be
+  specified when implementing the trait. This type is now used in the signature
+  of the `select` method to represent the set of file descriptors to monitor.
 - The `system::virtual::Process::blocked_signals` and
   `system::virtual::Process::pending_signals` methods now return a reference to
   `system::virtual::Sigset` instead of a reference to